src/installer/templates/skills/mistral/mistral-vision/requirements.txt
mistralai>=1.0.0
Pillow>=10.0.0
pyzbar>=0.1.9  # Für Barcode/QR-Code (optional)
httpx>=0.25.0